fix(test): keep primary entry below architecture gate
